==> jupyter/1.data_process.ipynb <==
# Data Clean

import re
import sys
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns
import string
import nltk
import warnings 
warnings.filterwarnings("ignore", category=DeprecationWarning)

sys.path.append("../") 
from personal_library.NLP.preprocess import (
    remove_pattern,
    rm_pun_num_esp_cha,
    rm_length_word,
    tokenize,
    stemmer,
    join_tokenize,
    hashtag_extract,
    count_caps,
    hashtag_rm
)

from personal_library.NLP.data_analysis import(
    plot_labels_wordcloud,
    plot_hashtag_hist,
)

from personal_library.NLP.core.model_preprocessors import (
    corpus2vec,
    standard_word2vec_size
)

from gensim.models import KeyedVectors
from sklearn.manifold import TSNE

%load_ext autoreload
%autoreload 2
%matplotlib inline

# Path to train and test files
train_path = '../data/train_E6oV3lV.csv'
test_path = '../data/test_tweets_anuFYb8.csv'

train  = pd.read_csv(train_path)
test = pd.read_csv(test_path)

all_data = train.append(test, ignore_index=True, sort=True)

train.head()

all_data['tidy_tweet'] = np.vectorize(remove_pattern)(all_data['tweet'], "@[\w]*")
all_data['tidy_tweet'] = rm_pun_num_esp_cha(all_data['tidy_tweet'])
all_data['tidy_tweet'] = rm_length_word(all_data['tidy_tweet'])
tokenized_tweet = tokenize(all_data['tidy_tweet'])
# tokenized_tweet = stemmer(tokenized_tweet)
all_data['tidy_tweet'] = join_tokenize(tokenized_tweet)
all_data['hashtag'] = hashtag_extract(all_data['tidy_tweet'], flatten=False)
all_data['tidy_tweet'] = np.vectorize(remove_pattern)(all_data['tidy_tweet'], "#[\w]*")

tokenized_tweet = tokenize(all_data['tidy_tweet'])
all_data.fillna('test', inplace = True)
all_data.to_csv('../data/pandas_data_frame.csv')
all_data["Name Length"] = all_data['tidy_tweet'].str.len()
all_data.head()

all_data.tail()

print("Median: ", all_data["Name Length"].median())
print("Media: ", all_data["Name Length"].mean())
print("Std: ", all_data["Name Length"].std())
print("MAx: ", all_data["Name Length"].max())

np.count_nonzero(all_data['label'] == 0.0)

# Word2Vec corpus Transformation

# Load word2vec Model
en_model = KeyedVectors.load_word2vec_format('../data/vectors/cc.en.300.vec')

import timeit

start = timeit.timeit()

test = corpus2vec(en_model, tokenized_tweet, use_next=True, debug=True)

end = timeit.timeit()
print(end - start)

np.save('../data/numpy_vectorize_data', test)
print(test.shape)

### Clean data and make dataset of same length, use MAX

%reload_ext autoreload

data = np.load('../data/numpy_vectorize_data.npy')
clean_data = standard_word2vec_size(data, 300)

# Clean and save train and test data

test = clean_data[all_data['label']=='test']
y = all_data['label'][(all_data['label']==1) | (all_data['label']==0)]
X = clean_data[(all_data['label']==1) | (all_data['label']==0)]

np.save('../data/dataset/X', X)
np.save('../data/dataset/y', y)
np.save('../data/dataset/test', test)

print(y.shape)
print(X.shape)
