==> jupyter/10.submission.ipynb <==
import re
import sys
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import seaborn as sns
import string
import nltk
import warnings 
warnings.filterwarnings("ignore", category=DeprecationWarning)

sys.path.append("../") 
from personal_library.NLP.preprocess import (
    remove_pattern,
    rm_pun_num_esp_cha,
    rm_length_word,
    tokenize,
    stemmer,
    join_tokenize,
    hashtag_extract,
    count_caps,
    hashtag_rm
)

from personal_library.NLP.data_analysis import(
    plot_labels_wordcloud,
    plot_hashtag_hist,
)

from personal_library.NLP.core.model_preprocessors import (
    corpus2vec,
    standard_word2vec_size
)

from gensim.models import KeyedVectors
from sklearn.manifold import TSNE

# Path to test file
test_path = '../data/test_tweets_anuFYb8.csv'
test = pd.read_csv(test_path)
test.head()

test['tidy_tweet'] = np.vectorize(remove_pattern)(test['tweet'], "@[\w]*")
test['tidy_tweet'] = rm_pun_num_esp_cha(test['tidy_tweet'])
test['tidy_tweet'] = rm_length_word(test['tidy_tweet'])
tokenized_tweet = tokenize(test['tidy_tweet'])
# tokenized_tweet = stemmer(tokenized_tweet)
test['tidy_tweet'] = join_tokenize(tokenized_tweet)
test['hashtag'] = hashtag_extract(test['tidy_tweet'], flatten=False)
test['tidy_tweet'] = np.vectorize(remove_pattern)(test['tidy_tweet'], "#[\w]*")

tokenized_tweet = tokenize(test['tidy_tweet'])
test.fillna('test', inplace = True)
test["Name Length"] = test['tidy_tweet'].str.len()
test.head()

print("Median: ", test["Name Length"].median())
print("Media: ", test["Name Length"].mean())
print("Std: ", test["Name Length"].std())
print("MAx: ", test["Name Length"].max())

import pickle
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

# loading
with open('../model_wehigts/tokenizer.pickle', 'rb') as handle:
    tk = pickle.load(handle)

max_len = 120
train_tokenized = tk.texts_to_sequences(test['tidy_tweet'])
x_test = pad_sequences(train_tokenized, maxlen=max_len)

print(x_test.shape)
